Improve Context classes and TreeWalker
- Don't use method_missing to bubble up method calls
- Record statement counting and call recievers everywhere
- Make RootContext know about its AST
- Move #default_assignments method to MethodContext
- Use full code example for config_for spec
- Limit TreeWalker's interface to #walk
- Remove MethodParameters module
