+++ lib/reek/context/root_context.rb
@@ -1,25 +1,14 @@
+require_relative 'code_context'
+
 module Reek
-  # @api private
   module Context
     #
     # A context wrapper representing the root of an abstract syntax tree.
     #
     # @api private
-    class RootContext
-      def initialize
-        @name = ''
-      end
-
-      def method_missing(_method, *_args)
-        nil
-      end
-
-      def config_for(_)
-        {}
-      end
-
-      def count_statements(_num)
-        0
+    class RootContext < CodeContext
+      def initialize(exp)
+        super(nil, exp)
       end
 
       def full_name
