+++ spec/reek/context/singleton_method_context_spec.rb
@@ -1,11 +1,10 @@
 require_relative '../../spec_helper'
 require_relative '../../../lib/reek/context/singleton_method_context'
-require_relative '../../../lib/reek/context/root_context'
 
 RSpec.describe Reek::Context::SingletonMethodContext do
   let(:smc) do
     sexp = s(:def, :foo, s(:args, s(:arg, :bar)), nil)
-    Reek::Context::SingletonMethodContext.new(Reek::Context::RootContext.new, sexp)
+    Reek::Context::SingletonMethodContext.new(nil, sexp)
   end
 
   describe '#envious_receivers' do
